♻️ refector roomlist to let it show avatar
