+++ app/components/room-chat/room-chat.tsx
@@ -23,8 +23,9 @@ import { useIsMobile } from "~/hooks/use-mobile";
 import { useRoomMessages } from "~/hooks/use-room-messages";
 import { client } from "~/lib/matrix-api/client";
 import { getRoom, getRoomTopic } from "~/lib/matrix-api/room";
-import { getRoomAvatar } from "~/lib/matrix-api/utils";
 import { InviteUserDialog } from "./invite-user-dialog";
+import { useRoomAvatar } from "~/hooks/use-room-avatar";
+import { avatarFallback } from "~/lib/utils";
 
 interface RoomChatProps {
   readonly onBackClick?: () => void;
@@ -35,6 +36,8 @@ export const RoomChat = memo(({ onBackClick = () => {} }: RoomChatProps) => {
   const isMobile = useIsMobile();
   const { selectedRoomId } = useRoomContext();
   const [room, setRoom] = useState(getRoom(selectedRoomId));
+
+  const roomAvatarUrl = useRoomAvatar(room);
 
   // selected room changes
   useEffect(() => {
@@ -80,16 +83,8 @@ export const RoomChat = memo(({ onBackClick = () => {} }: RoomChatProps) => {
             </Button>
           ) : null}
           <Avatar>
-            <AvatarImage
-              src={getRoomAvatar(room, room.client.baseUrl)}
-              alt={room?.name || ""}
-            />
-            <AvatarFallback>
-              {(room?.name || "")
-                .split(" ")
-                .map((n) => n[0])
-                .join("")}
-            </AvatarFallback>
+            <AvatarImage src={roomAvatarUrl} alt={room?.name || ""} />
+            <AvatarFallback>{avatarFallback(room.name)}</AvatarFallback>
           </Avatar>
           <div className="flex flex-row gap-2">
             <div>
