+++ app/routes/_home.home._index.tsx
@@ -1,7 +1,7 @@
 import { useEffect } from "react";
 import { useOutletContext } from "react-router";
 import { RoomChat } from "~/components/room-chat/room-chat";
-import { RoomList } from "~/components/room-list";
+import { RoomList } from "~/components/room-list/room-list";
 import {
   ResizableHandle,
   ResizablePanel,
